Build multiple target
