+++ packages/web-ui/webpack.config.node.js
@@ -1,0 +1,13 @@
+const _ = require('lodash');
+const commonConfig = require('./webpack.config.common');
+
+module.exports = _.merge({},
+  commonConfig,
+  {
+    target: 'node',
+    output: {
+      filename: 'index.node.js',
+      libraryTarget: 'commonjs2',
+    },
+  },
+);
